Pagar al contacto en vez del usuario

+++ app/src/main/java/com/example/zappay/request/ContactoRequest.kt
@@ -4,5 +4,5 @@ data class ContactoRequest(
     val Nombre: String,
     val Rut: String,
     val Correo: String,
-    var Saldo: Int
+    var Saldo: Double
 )
